### 2025-02-17 #7784 @warrickball
Hi @YuiKasagi — Thanks for your submission! All the suitable JOSS editors are currently working at capacity so I'm going to "waitlist" this review until an editor with the relevant expertise is available to take it on. Thanks for your patience!
