### 2025-04-15 6d1d7e2
fix ncw uncertainties
