### 2025-06-21 1c98fab
adopt npix to functions and set aperture width for IRCS
