### 2025-09-13 #7784 @YuiKasagi
Thank you very much, @arfon !
Based on the JOSS reviewers list and the review histories of past JOSS papers, I suggest the following potential reviewers:
- Wilfred Tyler Gee (@\wtgee)
- R. THOMAS (@\astrom-tom)
- Gabriel Perren (@\Gabriel-p)
- Carter Lee Rhea (@\crhea93)
- Arjun Savel (@\arjunsavel)
