### 2025-10-01 #9126 @editorialbot
[REVIEW]: PyIRD: A Python-Based Data Reduction Pipeline for Subaru/IRD and REACH
**Submitting author:** @YuiKasagi (<a href="http://orcid.org/0000-0002-8607-358X">Yui Kasagi</a>)
**Repository:** https://github.com/prvjapan/pyird
**Branch with paper.md** (empty if default branch): master
**Version:** v1.2.0
