### 2025-10-10 prvjapan/pyird#120 @crhea93
Issues running tests
I am having trouble running the unit and integration tests. Could you please add instructions on how to do this?

Since there aren't any instructions (at least I was unable to find them), I tried running `pytest` from both the root and the test directories. However, both time I received the following error messages:

```shell
___________________________________________________________________ ERROR collecting tests/unittests/image/test_aptrace.py ___________________________________________________________________
import file mismatch:
imported module 'test_aptrace' has this __file__ attribute:
  /home/carterrhea/Documents/pyird/tests/integration/image/test_aptrace.py
which is not the same as the test file we want to collect:
  /home/carterrhea/Documents/pyird/tests/unittests/image/test_aptrace.py
HINT: remove __pycache__ / .pyc files and/or use a unique basename for your test file modules
____________________________________________________________________ ERROR collecting tests/unittests/image/test_bias.py _____________________________________________________________________
import file mismatch:
imported module 'test_bias' has this __file__ attribute:
  /home/carterrhea/Documents/pyird/tests/integration/image/test_bias.py
which is not the same as the test file we want to collect:
  /home/carterrhea/Documents/pyird/tests/unittests/image/test_bias.py
HINT: remove __pycache__ / .pyc files and/or use a unique basename for your test file modules
____________________________________________________________________ ERROR collecting tests/unittests/spec/test_wavcal.py ____________________________________________________________________
import file mismatch:
imported module 'test_wavcal' has this __file__ attribute:
  /home/carterrhea/Documents/pyird/tests/integration/spec/test_wavcal.py
which is not the same as the test file we want to collect:
  /home/carterrhea/Documents/pyird/tests/unittests/spec/test_wavcal.py
HINT: remove __pycache__ / .pyc files and/or use a unique basename for your test file modules
```

This is linked to JOSS issue #9126 https://github.com/openjournals/joss-reviews/issues/9126
