### 2025-10-10 prvjapan/pyird#121 @crhea93
Hotpixel Identification
Looking at the documentation, there appear to be two well--documented options for removing hotpixels: `pyird.image.hotpix.identify_hotpix` and `pyird.image.hotpix.identify_hotpix_sigclip`. Have these been benchmarked against started hot pixel identification algorithms such as [LACosmic](https://lacosmic.readthedocs.io/en/stable/)?

This is related to the JOSS issue #9126 https://github.com/openjournals/joss-reviews/issues/9126
