### 2025-10-10 prvjapan/pyird#122 @crhea93
Update docstrings to include more details on implementation
The docstrings contain a considerably amount of detail on the function arguments and what they return. However, I think it would be useful if they, in general, contained more information on algorithmic implementations. Since this package has many API docstrings, it is not feasible to do this for every docstring. Instead, I would suggest concentrating on the higher level API calls and more important functions such as ` pyird.spec.uncertainty.FluxUncertainty` and ` pyird.spec.continuum.ContinuumFit`.

This is related to the JOSS issue #9126 https://github.com/openjournals/joss-reviews/issues/9126
