### 2025-10-14 c787e9d
Add instruction to run pytest
