+++ documents/developers/pytest.rst
@@ -1,0 +1,33 @@
+Test code for developers
+==========================
+
+PyIRD includes test code to ensure the maintainability of the package, using the `pytest <https://docs.pytest.org/en/stable/>`_ testing framework.
+
+How to run tests
+^^^^^^^^^^^^^^^^^
+To run tests, execute the following command in the root directory of the repository:
+
+.. code-block:: bash
+
+   pytest tests/unittests
+   pytest tests/integration
+
+or, 
+
+.. code-block:: bash
+
+   cd tests/unittests
+   pytest
+
+Note that tests in ``tests/unittests`` are automatically executed by GitHub Actions when a pull request is made to the develop or master branch.
+
+Unit tests
+""""""""""""
+
+``tests/unittests``: Tests for individual functions and classes. 
+Developers encauraged to create new tests when adding new features.
+
+Integration tests
+"""""""""""""""""""
+
+``tests/integration``: Tests for the behavior of multiple functions and classes working together.
