### 2025-10-26 prvjapan/pyird#121 @YuiKasagi
Thank you for your comment. I had not compared them before, so I did a quick test and summarized the results below.

I checked the differences between the results obtained from `pyird.image.hotpix.identify_hotpix`, `pyird.image.hotpix.identify_hotpix_sigclip`, and LACosmic.
(I installed LACosmic from https://github.com/larrybradley/lacosmic.)

> Parameter settings:
> `identify_hotpix`: default parameters
> `identify_hotpix_sigclip`: default parameters
> `lacosmic`: `contrast=1.`, `cr_threshold=6.`, `neighbor_threshold=3.` (and `effective_gain`, `readnoise` of [IRD](https://ird.mtk.nao.ac.jp/IRDpub/index_tmp.html))

The histograms below compare the pixel value distributions between the original dark image and the masked image for the Y/J-band detector (left) and the H-band detector (right).
I found that LACosmic (shown as the filled orange histogram) and `identify_hotpix_sigclip` from PyIRD (shown as the green step-line histogram) perform at almost the same level.
For the H-band image, `identify_hotpix_sigclip` appears to mask hot pixels more strictly. I tried adjusting several parameters of LACosmic, but the masking level did not change significantly beyond that.
On the other hand, `identify_hotpix` (shown as the red step-line histogram) does not work as well, likely because it uses [SEP](https://sep.readthedocs.io/en/stable/index.html), which is designed for detecting point sources rather than hot pixels.

IRD images also contain some bad pixels with negative values, and I found that LACosmic cannot detect them.
The zoomed images of the original Y-band image below show the hot pixel positions marked with pink rectangles.
There are several hot pixels missed by LACosmic (for example, part of the cross-pattern hot pixels in the lower-left region), including some negative-value pixels (e.g., around (x, y) = (523, 534)).
I think the advantages of `identify_hotpix_sigclip` are: (1) it can also detect bad pixels with negative values, and (2) its parameter `frac` automatically controls the masking process to prevent excessive masking.

I am planning to add a link to this issue in the API documentation so that these results can serve as guidance for using the method.
