### 2025-10-27 e27f309
add instruction for fitsid
