### 2025-10-30 2783b3c
update output messages
