### 2025-10-30 2c7376c
nap to num_aperture
