+++ src/pyird/image/aptrace.py
@@ -7,13 +7,13 @@ import warnings
 from pyird.plot.order import plot_crosssection, plot_tracelines
 
 
-def cross_section(dat, nrow, nap):
+def cross_section(dat, nrow, num_aperture):
     """extract cross section in row direction and search peaks.
 
     Args:
         dat: flat data
         nrow: row number to be extracted
-        nap: number of total apertures to be traced
+        num_aperture: number of total apertures to be traced
 
     Returns:
         masked data and peaks at the cross section
@@ -24,7 +24,7 @@ def cross_section(dat, nrow, nap):
     i = 0
     diffstd = 100
     peakind = []
-    while (len(peakind) < nap or (diffstd > 10)) and (i < len(heights)):
+    while (len(peakind) < num_aperture or (diffstd > 10)) and (i < len(heights)):
         height = heights[i]
         peakind, _ = find_peaks(onerow, height=height, distance=5)
         diffstd = np.std(np.diff(peakind)[::2])
@@ -43,65 +43,66 @@ def cross_section(dat, nrow, nap):
     return onerow_masked, peakind_cut
 
 
-def set_aperture(dat, cutrow, nap, ign_ord=[], plot=True):
+def set_aperture(dat, search_start_row, num_aperture, ign_ord=[], plot=True):
     """search aperture
 
     Args:
         dat: flat data
-        cutrow: row number used to set aperture
-        nap: number of total apertures to be traced
+        search_start_row: starting row number to search apertures
+        num_aperture: number of total apertures to be traced
         ign_ord: orders to be ignored
         plot: show figure of selected apertures
 
     Returns:
-        peak position in the cross section at cutrow
+        peak position in the cross section at search_row
     """
     if len(ign_ord) >= 30:
        raise ValueError("length of ign_ord should be < 30.")
 
-    # Search peaks in the cross section at cutrow
-    # cutrow is selected so that the number of peaks and nap match
-    cutrow_min = 500
-    cutrow_max = 1550
+    # Search peaks in the cross section at search_row
+    # search_row is selected so that the number of peaks and num_aperture match
+    search_row_min = 500
+    search_row_max = 1550
     peakind_cut = []
-    cutrow_lim = True
+    search_row_lim = True
     prange = False
-    while ((len(peakind_cut) != nap) or prange) and cutrow_lim:
-        onerow_masked, peakind_cut = cross_section(dat, cutrow, nap)
+    search_row = search_start_row
+    while ((len(peakind_cut) != num_aperture) or prange) and search_row_lim:
+        onerow_masked, peakind_cut = cross_section(dat, search_row, num_aperture)
         # mask peaks for selected apertures
         if len(ign_ord)>0:
             mask = np.ones_like(peakind_cut, dtype=bool)
             mask[np.array(ign_ord)-1] = False
             peakind_cut = np.array(peakind_cut)[mask]
-        if nap in [21, 42]:  # h band
+        if num_aperture in [21, 42]:  # h band
             prange = (peakind_cut[-1] > 1500) or (peakind_cut[0] > 40)
-        elif nap in [51, 102]:  # yj band
+        elif num_aperture in [51, 102]:  # yj band
             prange = (peakind_cut[0] < 250) or (peakind_cut[-1] < 2000)
-        cutrow_lim = (cutrow > cutrow_min) and (cutrow < cutrow_max)
-        cutrow += 1
-    if not cutrow_lim:
+        search_row_lim = (search_row > search_row_min) and (search_row < search_row_max)
+        search_row += 1
+    if not search_row_lim:
         warnings.warn(
-            'Please change "cutrow" or "nap" or confirm the orientation of the image.',
+            'Please change "search_row" or "num_aperture" or confirm the orientation of the image.',
[…]
