+++ documents/tutorials/IRD_stream.rst
@@ -5,9 +5,10 @@ This tutorial demonstrates how to reduce raw data to
 wavelength-calibrated 1D spectra. By using the ``Stream2D`` framework,
 you can apply functions to multiple FITS files efficiently.
 
-The same pipeline is available in ``examples/python/IRD_stream.py``. For
-REACH data, refer the script available in
-``examples/python/REACH_stream.py``.
+The same pipeline is available in
+`examples/python/IRD_stream.py <https://github.com/prvjapan/pyird/blob/master/examples/python/IRD_stream.py>`_.
+For REACH data, refer the script available in
+`examples/python/REACH_stream.py <https://github.com/prvjapan/pyird/blob/master/examples/python/REACH_stream.py>`_.
 
 - :ref:`step0`
 - :ref:`step1`
@@ -33,7 +34,8 @@ Step 0: Settings
 Directory Structure
 ~~~~~~~~~~~~~~~~~~~
 
-First, specify the path to the raw data directory (``datadir``) and the output directory (``anadir``).
+First, specify the path to the raw data directory (``datadir``) and the
+output directory (``anadir``)
 
 This tutorial assumes the following directory structure:
 
@@ -51,7 +53,8 @@ This tutorial assumes the following directory structure:
 
 In this structure, the flat, thar, target, and dark directories are part
 of the ``datadir``, each containing raw data for ‘Flat’, ‘ThAr’,
-‘Target’, and optionally ‘Dark’ frames. The reduc directory is used as
+‘Target’, and optionally ‘Dark’ frames (see also :doc:`../input_data`).
+The reduc directory is used as
 ``anadir`` for storing processed data.
 
 .. code:: ipython3
@@ -69,8 +72,8 @@ Specify the Data to be Analyzed
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 Please change the following variables based on the data you want to
-analyze.
-The sample data can be downloaded from the `Zenodo repository <https://zenodo.org/records/14614004>`_
+analyze. The sample data can be downloaded from the
+`Zenodo repository <https://zenodo.org/records/14614004>`_
 
 .. code:: ipython3
 
@@ -78,36 +81,41 @@ The sample data can be downloaded from the `Zenodo repository <https://zenodo.or
     mmf = 'mmf2' #'mmf1' (comb fiber) or 'mmf2' (star fiber)
     readout_noise_mode = 'default'
 
-**Note**:
+.. note::
 
-Ensure that the ``readout_noise_mode`` is set to either ‘real’ or
-‘default’.
+  Ensure that the ``readout_noise_mode`` is set to either ‘real’ or
+  ‘default’.
 
-- ``readout_noise_mode = 'real'``: Need to reduce the dataset with
-  ``band = 'y'`` and ``mmf = 'mmf1'`` at first.
+  - ``readout_noise_mode = 'real'``: Need to reduce the dataset with
+    ``band = 'y'`` and ``mmf = 'mmf1'`` at first.
 
-  - With this setting, uncertainties and signal-to-noise ratio at each
-    wavelength will be included in the output files (**nw…_m?.dat**
-    and **ncw…_m?.dat**).
-  - Those values are based on the readout noise (RN) calculated using
-    the comb spectrum (in mmf1) of the Y/J band.
+    - With this setting, uncertainties and signal-to-noise ratio at each
+      wavelength will be included in the output files (**nw…\_m?.dat**
+      and **ncw…\_m?.dat**).
+    - Those values are based on the readout noise (RN) calculated using
+      the comb spectrum (in mmf1) of the Y/J band.
 
[…]
