+++ documents/tutorials/plot_spec.rst
@@ -1,6 +1,25 @@
 Output Spectra
 ==============
 
+When running :doc:``./IRD_stream``, three types of 1D spectra are
+generated:
+
+- Wavelength calibrated 1D spectrum (\*\*’w…\_m?.dat’\*\*) with columns:
+  ``$1: Wavelength [nm]``, ``$2: Order``, ``$3: Counts``.
+- Normalized 1D spectrum (\*\*’nw…\_m?.dat’\*\*) with columns:
+  ``$1: Wavelength [nm]``, ``$2: Order``, ``$3: Counts``, ``$4: S/N``,
+  ``$5: Uncertainties``
+- Order-combined normalized 1D spectrum (\*\*’ncw…\_m?.dat’\*\*) with
+  columns: ``$1: Wavelength [nm]``, ``$2: Counts``, ``$3: S/N``,
+  ``$4: Uncertainties``
+
+.. note::
+
+::
+
+   - The normalized spectrum (nw) is obtained by dividing the extracted spectrum (w) by the blaze function (`wblaze_?_m?.dat`).
+   - The order-combined spectrum (ncw) demonstrates an improved S/N in the wavelength region where adjacent orders overlap.
+
 .. code:: ipython3
 
     import pathlib
@@ -28,22 +47,18 @@ Output Spectra
 
     <>:10: SyntaxWarning: invalid escape sequence '\s'
     <>:10: SyntaxWarning: invalid escape sequence '\s'
-    /var/folders/sb/rxwbk6kd4gldqpdvzxnhq4xm0000gn/T/ipykernel_1300/3404868316.py:10: SyntaxWarning: invalid escape sequence '\s'
+    /var/folders/sb/rxwbk6kd4gldqpdvzxnhq4xm0000gn/T/ipykernel_11979/3404868316.py:10: SyntaxWarning: invalid escape sequence '\s'
       readargs = {"header": None, "sep": "\s+"}
 
 
 The figure shows the target spectra extracted using ``PyIRD``.
 
-- The sample dataset corresponds to the brown dwarf G196-3B.
+.. note::
 
-- The emission-like signals observed in the spectra are likely due to
-  sky (airglow) emissions or hotpixels that were not masked.
+::
 
-- The normalized spectrum (nw) is obtained by dividing the extracted
-  spectrum (w) by the blaze function (``wblaze_?_m?.dat``).
-
-- The order-combined spectrum (ncw) demonstrates an improved S/N in the
-  wavelength region where adjacent orders overlap.
+   - The sample dataset corresponds to the brown dwarf G196-3B.
+   - The emission-like signals observed in the spectra are likely due to sky (airglow) emissions or hotpixels that were not masked.
 
 .. code:: ipython3
 
@@ -58,7 +73,7 @@ The figure shows the target spectra extracted using ``PyIRD``.
         nwfile_path = anadir / f"nw{fitsid}_{mmf[0]}{mmf[-1]}.dat"
         nwspec = pd.read_csv(nwfile_path, names = names, **readargs)
     
-        # Order-combined Spectrum
+        # Order-combined Spectrum:
         # $1: wavelength [nm], $2: counts, $3: sn_ratio, $4: uncertainty
         ncwfile_path = anadir / f"ncw{fitsid}_{mmf[0]}{mmf[-1]}.dat"
         names_ncw = [x for i, x in enumerate(names) if i != 1]
