+++ paper/paper.bib
@@ -59,6 +59,162 @@
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@misc{iraf_community,
+  title = {IRAF Community Distribution},
+  howpublished = {\url{https://iraf-community.github.io/}},
+  note = {Accessed: 2025-11-02}
+}
+
+@article{stsci2018,
+  title={Removing the Institute’s Dependence on IRAF: You Can Do It Too},
+  journal={STScI Newsletter},
+  volume={35},
+  number={03},
+  year={2018},
+  url={https://www.stsci.edu/contents/newsletters/2018-volume-35-issue-03/removing-the-institutes-dependence-on-iraf-you-can-do-it-too}
+}
+
+@ARTICLE{Hamano2024,
+       author = {{Hamano}, Satoshi and {Ikeda}, Yuji and {Otsubo}, Shogo and {Katoh}, Haruki and {Fukue}, Kei and {Matsunaga}, Noriyuki and {Taniguchi}, Daisuke and {Kawakita}, Hideyo and {Takenaka}, Keiichi and {Kondo}, Sohei and {Sameshima}, Hiroaki},
+        title = "{WARP: The Data Reduction Pipeline for the WINERED Spectrograph}",
+      journal = {\pasp},
+     keywords = {Spectroscopy, Near infrared astronomy, Astronomy data reduction, 1558, 1093, 1861, Astrophysics - Instrumentation and Methods for Astrophysics},
+         year = 2024,
+        month = jan,
+       volume = {136},
+       number = {1},
+          eid = {014504},
+        pages = {014504},
+          doi = {10.1088/1538-3873/ad1b38},
+archivePrefix = {arXiv},
+       eprint = {2401.04876},
+ primaryClass = {astro-ph.IM},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2024PASP..136a4504H},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{Sim2014,
+       author = {{Sim}, Chae Kyung and {Le}, Huynh Anh Nguyen and {Pak}, Soojong and {Lee}, Hye-In and {Kang}, Wonseok and {Chun}, Moo-Young and {Jeong}, Ueejeong and {Yuk}, In-Soo and {Kim}, Kang-Min and {Park}, Chan and {Pavel}, Michael D. and {Jaffe}, Daniel T.},
+        title = "{Comprehensive data reduction package for the Immersion GRating INfrared Spectrograph: IGRINS}",
+      journal = {Advances in Space Research},
+         year = 2014,
+        month = jun,
+       volume = {53},
+       number = {11},
+        pages = {1647-1656},
+          doi = {10.1016/j.asr.2014.02.024},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2014AdSpR..53.1647S},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@article{pypeit:joss_pub,
+    doi = {10.21105/joss.02308},
+    url = {https://doi.org/10.21105/joss.02308},
+    year = {2020},
+    publisher = {The Open Journal},
+    volume = {5},
+    number = {56},
+    pages = {2308},
+    author = {J. Xavier Prochaska and Joseph F. Hennawi and Kyle B. Westfall and Ryan J. Cooke and Feige Wang and Tiffany Hsyu and Frederick B. Davies and Emanuele Paolo Farina and Debora Pelliccia},
+    title = {PypeIt: The Python Spectroscopic Data Reduction Pipeline},
[…]
