+++ paper/paper.bib
@@ -59,21 +59,6 @@
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@misc{iraf_community,
-  title = {IRAF Community Distribution},
-  howpublished = {\url{https://iraf-community.github.io/}},
-  note = {Accessed: 2025-11-02}
-}
-
-@article{stsci2018,
-  title={Removing the Institute’s Dependence on IRAF: You Can Do It Too},
-  journal={STScI Newsletter},
-  volume={35},
-  number={03},
-  year={2018},
-  url={https://www.stsci.edu/contents/newsletters/2018-volume-35-issue-03/removing-the-institutes-dependence-on-iraf-you-can-do-it-too}
-}
-
 @ARTICLE{Hamano2024,
        author = {{Hamano}, Satoshi and {Ikeda}, Yuji and {Otsubo}, Shogo and {Katoh}, Haruki and {Fukue}, Kei and {Matsunaga}, Noriyuki and {Taniguchi}, Daisuke and {Kawakita}, Hideyo and {Takenaka}, Keiichi and {Kondo}, Sohei and {Sameshima}, Hiroaki},
         title = "{WARP: The Data Reduction Pipeline for the WINERED Spectrograph}",
