+++ documents/userguide/backgrounds.rst
@@ -4,6 +4,6 @@ Background
 .. toctree::
    :maxdepth: 1
    :caption: Background and Principles of Pipeline Design
-	     
-   userguide/pattern.rst
-   userguide/terms.rst
+
+   pattern.rst
+   terms.rst
