+++ documents/developers.rst
@@ -4,5 +4,5 @@ For Developers
 .. toctree::
    :maxdepth: 1
    :caption: For developers
-	     
-    developers/pytest.rst
+
+   developers/pytest
