### 2025-11-20 #9126 @YuiKasagi
Since I believe I have addressed all reviewer issues raised so far and there has been no new feedback recently, I have gone ahead and merged the `joss-review` branch into `master` for now. I have also updated the [documentation](https://secondearths.sakura.ne.jp/pyird/) accordingly.

If there are any additional comments or issues, please let me know and I will be happy to address them.
