+++ paper/paper.md
@@ -23,10 +23,11 @@ authors:
     orcid: 0000-0002-4677-9182
     affiliation: "3, 5"
   - name: Kento Masuda
+    orcid: 0000-0003-1298-9699
     affiliation: 6
 
 affiliations:
- - name: Institute of Space and Astronomical Science, Japan Aerospace Exploration Agency, 3-1-1 Yoshinodai, Chuo-ku, Sagamihara, Kanagawa, 252-5210, Japan
+ - name: Institute of Space and Astronoutical Science, Japan Aerospace Exploration Agency, 3-1-1 Yoshinodai, Chuo-ku, Sagamihara, Kanagawa, 252-5210, Japan
    index: 1
  - name: Department of Astronomy, Graduate School of Science, The University of Tokyo, 7-3-1 Hongo, Bunkyo-ku, Tokyo 113-0033, Japan
    index: 2
@@ -36,9 +37,9 @@ affiliations:
    index: 4
  - name: National Astronomical Observatory of Japan, 2-21-1 Osawa, Mitaka, Tokyo 181-8588, Japan
    index: 5
- - name: Department of Earth and Space Science, Osaka University, Osaka 560-0043, Japan
+ - name: Department of Earth and Space Science, Osaka University, Toyonaka, Osaka 560-0043, Japan
    index: 6
-date: 12 December 2024
+date: 13 January 2026
 bibliography: paper.bib
 
 # Optional fields if submitting to a AAS journal too, see this blog post:
@@ -59,7 +60,7 @@ Additionally, the pipeline includes an updated method for removing readout noise
 
 # Statement of need
 
-The reduction of high-dispersion spectroscopic data has traditionally been performed using `IRAF`, one of the most widely used software tools for astronomical data reduction and analysis.
+The reduction of high-dispersion spectroscopic data has traditionally been performed using `IRAF`, one of the most widely used tools for astronomical data reduction and analysis.
 Although the National Optical Astronomy Observatory (NOAO) officially ceased development and maintenance of `IRAF` in 2013, community-based maintenance has continued. 
 However, the official IRAF community distribution[^iraf_community] and the Space Telescope Science Institute (STScI)[^stsci] have both recommended that researchers transition away from `IRAF` due to its legacy architecture and lack of institutional support.
 
@@ -68,12 +69,12 @@ However, the official IRAF community distribution[^iraf_community] and the Space
 
 In recent years, several open-source, Python-based pipelines for the reduction of near-infrared echelle spectrographs have been developed.
 Some pipelines utilize `PyRAF`, a Python interface to IRAF, such as `WARP` for the WINERED spectrograph [@Hamano2024], while others, including `PLP` for IGRINS [@Sim2014] and `PypeIt` [@pypeit:joss_pub], do not rely on PyRAF or IRAF-based components.
-While these pipelines provide general frameworks or instrument-specific solutions, `PyIRD` is designed to offer a simple pipeline optimized for IRD and REACH data reduction.
+While these pipelines provide either general frameworks or instrument-specific solutions, `PyIRD` is designed to offer a simple pipeline optimized for IRD and REACH data reduction.
 Furthermore, recent advances combining adaptive optics with these instruments have enabled high-resolution spectroscopic observations of faint companions orbiting bright main-sequence stars.
 To support such observations, `PyIRD` implements improved detector noise reduction to extract high-quality spectra from these faint targets.
 
 Together, these developments underscore the need for actively maintained, scalable, and flexible software for high-resolution spectroscopic data reduction.
[…]
