+++ src/pyird/__init__.py
@@ -1,6 +1,6 @@
 __all__ = ['fitsset', 'irdstream']
 
-__version__ = '1.1.0'
+__version__ = '1.2.0'
 
 #from . import fitsset
 #from . import irdstream
