### 2026-01-14 6fef7c5
update pyproject.toml and setup.cfg
