+++ src/pyird/__init__.py
@@ -1,11 +1,3 @@
 __all__ = ['fitsset', 'irdstream']
 
 __version__ = '1.2.0'
-
-#from . import fitsset
-#from . import irdstream
-__url__ = 'https://github.com/prvjapan/pyird'
-__author__ = 'Yui Kasagi, Hajime Kawahara, and Collaborators'
-__email__ = 'kasagi.y96@gmail.com'
-__license__ = 'MIT'
-__description__ = 'Python pipeline for infrared detectors, and/or IRD/REACH'
