### 2026-01-16 #9126 @YuiKasagi
Thank you for the message, @arfon
Following your instructions:
- **PyIRD v1.2.0** has been released on GitHub, including all changes resulting from this review.
- The Zenodo archive DOI (version-specific) is: **10.5281/zenodo.18264940**
(Please note that the Zenodo record shows version 1.2.1 due to the timing of the GitHub–Zenodo integration; however, the contents are identical to v1.2.0.)

Thanks so much for your assistance and please let me know if anything else is needed.
